fix: move call system message to daemon, check max participants when creating a token to avoid giving tokens but erroring when attempting to join, check if the channel actually supports voice

+++ crates/core/database/src/voice/voice_client.rs
@@ -103,7 +103,7 @@ impl VoiceClient {
             .create_room(
                 channel.id(),
                 CreateRoomOptions {
-                    max_participants: voice.max_users.unwrap_or(0),
+                    max_participants: voice.max_users.unwrap_or(0) as u32,
                     empty_timeout: 5 * 60, // 5 minutes,
                     ..Default::default()
                 },
